+++ CHANGELOG.md
@@ -49,6 +49,7 @@
 - `/users/byUserId/{userId}` endpoint added as `Iterable::Users#for_id`
 - `/users/getFields` endpoint added as `Iterable::Users#fields`
 - `/users/updateEmail` endpoint added as `Iterable::Users#update_email`
+- `/users/{email}` delete endpoint added as `Iterable::Users#delete`
 
 ## v0.1.0 - 2017-10-04